Split makePickerWith state hoc
