+++ lib/.size-snapshot.json
@@ -1,26 +1,26 @@
 {
   "build/dist/material-ui-pickers.esm.js": {
-    "bundled": 143471,
-    "minified": 78671,
-    "gzipped": 21103,
+    "bundled": 148219,
+    "minified": 81418,
+    "gzipped": 21568,
     "treeshaked": {
       "rollup": {
-        "code": 64934,
-        "import_statements": 2099
+        "code": 66594,
+        "import_statements": 2167
       },
       "webpack": {
-        "code": 72410
+        "code": 74225
       }
     }
   },
   "build/dist/material-ui-pickers.umd.js": {
-    "bundled": 597704,
-    "minified": 222450,
-    "gzipped": 45688
+    "bundled": 602739,
+    "minified": 224747,
+    "gzipped": 46056
   },
   "build/dist/material-ui-pickers.umd.min.js": {
-    "bundled": 537438,
-    "minified": 203944,
-    "gzipped": 40910
+    "bundled": 542486,
+    "minified": 206234,
+    "gzipped": 41259
   }
 }
