+++ lib/src/Picker/WithViewsProps.tsx
@@ -1,0 +1,15 @@
+import { DateTimePickerView } from '../DateTimePicker';
+import { BasePickerProps } from '../typings/BasePicker';
+import { ExportedDateInputProps } from '../_shared/PureDateInput';
+import { DateValidationProps } from '../_helpers/text-field-helper';
+
+export interface WithViewsProps<T extends DateTimePickerView> {
+  /**
+   * Array of views to show
+   */
+  views?: T[];
+  /** First view to show */
+  openTo?: T;
+}
+
+export type WithDateInputProps = DateValidationProps & BasePickerProps & ExportedDateInputProps;
